+++ yacg/model/config.py
@@ -109,6 +109,9 @@ class Task:
         #: A task to run
         self.multiFileTask = None
 
+        #: A task to run
+        self.randomDataTask = None
+
     @classmethod
     def dictToObject(cls, dict):
         if dict is None:
@@ -132,6 +135,8 @@ class Task:
         obj.singleFileTask = SingleFileTask.dictToObject(dict.get('singleFileTask', None))
 
         obj.multiFileTask = MultiFileTask.dictToObject(dict.get('multiFileTask', None))
+
+        obj.randomDataTask = RandomDataTask.dictToObject(dict.get('randomDataTask', None))
         return obj
 
 
@@ -379,5 +384,223 @@ class MultiFileTaskFileFilterTypeEnum(Enum):
         else:
             return ''
 
+
+
+class RandomDataTask:
+    def __init__(self):
+
+        self.outputType = None
+
+        self.destDir = None
+
+        self.defaultMinElemCount = 1
+
+        self.defaultMaxElemCount = 10
+
+        self.specialElemCounts = []
+
+        self.defaultKeyPropNames = []
+
+        self.specialKeyPropNames = []
+
+        self.valuePools = []
+
+        self.defaultMinSize = None
+
+        self.defaultMaxSize = None
+
+        self.specialArraySizes = []
+
+        self.defaultMaxDepth = 2
+
+        self.specialMaxDepths = []
+
+    @classmethod
+    def dictToObject(cls, dict):
+        if dict is None:
+            return None
+        obj = cls()
+
+        obj.outputType = RandomDataTaskOutputTypeEnum.valueForString(dict.get('outputType', None))
+
+        obj.destDir = dict.get('destDir', None)
+
+        obj.defaultMinElemCount = dict.get('defaultMinElemCount', None)
+
+        obj.defaultMaxElemCount = dict.get('defaultMaxElemCount', None)
+
+        arraySpecialElemCounts = dict.get('specialElemCounts', [])
+        for elemSpecialElemCounts in arraySpecialElemCounts:
+            obj.specialElemCounts.append(
+                RandomDataTaskSpecialElemCounts.dictToObject(elemSpecialElemCounts))
+
+        arrayDefaultKeyPropNames = dict.get('defaultKeyPropNames', [])
+        for elemDefaultKeyPropNames in arrayDefaultKeyPropNames:
+            obj.defaultKeyPropNames.append(elemDefaultKeyPropNames)
+
+        arraySpecialKeyPropNames = dict.get('specialKeyPropNames', [])
+        for elemSpecialKeyPropNames in arraySpecialKeyPropNames:
+            obj.specialKeyPropNames.append(
+                RandomDataTaskSpecialKeyPropNames.dictToObject(elemSpecialKeyPropNames))
+
+        arrayValuePools = dict.get('valuePools', [])
+        for elemValuePools in arrayValuePools:
+            obj.valuePools.append(
+                RandomDataTaskValuePools.dictToObject(elemValuePools))
+
+        obj.defaultMinSize = dict.get('defaultMinSize', None)
+
+        obj.defaultMaxSize = dict.get('defaultMaxSize', None)
+
+        arraySpecialArraySizes = dict.get('specialArraySizes', [])
+        for elemSpecialArraySizes in arraySpecialArraySizes:
+            obj.specialArraySizes.append(
+                RandomDataTaskSpecialArraySizes.dictToObject(elemSpecialArraySizes))
+
+        obj.defaultMaxDepth = dict.get('defaultMaxDepth', None)
+
+        arraySpecialMaxDepths = dict.get('specialMaxDepths', [])
+        for elemSpecialMaxDepths in arraySpecialMaxDepths:
+            obj.specialMaxDepths.append(
+                RandomDataTaskSpecialMaxDepths.dictToObject(elemSpecialMaxDepths))
+        return obj
+
+
+class RandomDataTaskOutputTypeEnum(Enum):
+    JSON = 'JSON'
+    CSV = 'CSV'
+
+    @classmethod
+    def valueForString(cls, stringValue):
+        lowerStringValue = stringValue.lower() if stringValue is not None else None
+        if lowerStringValue is None:
+            return None
+        elif lowerStringValue == 'json':
+            return RandomDataTaskOutputTypeEnum.JSON
+        elif lowerStringValue == 'csv':
+            return RandomDataTaskOutputTypeEnum.CSV
+        else:
+            return None
+
+    @classmethod
+    def valueAsString(cls, enumValue):
+        if enumValue is None:
+            return ''
+        elif enumValue == RandomDataTaskOutputTypeEnum.JSON:
+            return 'JSON'
+        elif enumValue == RandomDataTaskOutputTypeEnum.CSV:
+            return 'CSV'
+        else:
+            return ''
+
+
+
+class RandomDataTaskSpecialElemCounts:
+    def __init__(self):
+
+        self.typeName = None
+
+        self.minElemCount = None
+
+        self.maxElemCount = None
+
+    @classmethod
+    def dictToObject(cls, dict):
+        if dict is None:
+            return None
+        obj = cls()
+
+        obj.typeName = dict.get('typeName', None)
+
+        obj.minElemCount = dict.get('minElemCount', None)
+
+        obj.maxElemCount = dict.get('maxElemCount', None)
+        return obj
+
+
+class RandomDataTaskSpecialKeyPropNames:
+    def __init__(self):
+
+        self.typeName = None
+
+        self.keyPropName = None
+
+    @classmethod
+    def dictToObject(cls, dict):
+        if dict is None:
+            return None
+        obj = cls()
+
+        obj.typeName = dict.get('typeName', None)
+
+        obj.keyPropName = dict.get('keyPropName', None)
+        return obj
+
+
+class RandomDataTaskValuePools:
+    def __init__(self):
+
+        self.propertyName = None
+
+        self.useAll = False
+
+        self.values = []
+
+    @classmethod
+    def dictToObject(cls, dict):
+        if dict is None:
+            return None
+        obj = cls()
+
+        obj.propertyName = dict.get('propertyName', None)
+
+        obj.useAll = dict.get('useAll', None)
+
+        arrayValues = dict.get('values', [])
+        for elemValues in arrayValues:
+            obj.values.append(elemValues)
+        return obj
+
+
+class RandomDataTaskSpecialArraySizes:
+    def __init__(self):
+
+        self.propertyName = None
+
+        self.minSize = None
+
+        self.maxSize = None
+
+    @classmethod
+    def dictToObject(cls, dict):
+        if dict is None:
+            return None
+        obj = cls()
+
+        obj.propertyName = dict.get('propertyName', None)
+
+        obj.minSize = dict.get('minSize', None)
+
+        obj.maxSize = dict.get('maxSize', None)
+        return obj
+
+
+class RandomDataTaskSpecialMaxDepths:
+    def __init__(self):
+
+        self.propertyName = None
+
+        self.maxDepth = None
+
+    @classmethod
+    def dictToObject(cls, dict):
+        if dict is None:
+            return None
+        obj = cls()
+
+        obj.propertyName = dict.get('propertyName', None)
+
+        obj.maxDepth = dict.get('maxDepth', None)
+        return obj