+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.16.0
+* enable generation of test data
+* rename property attrib 'implicitRef' to 'foreignKey'
+* introduce '__key' model extension, to mark properties as logical keys
+* introduce '__visualKey' model extension, to mark properties as visual key (e.g. name, label, ...) 
+
+# 0.15.0
+* add protobuffer template
+* add ordinal attribute to complex type properties
+
 # 0.14.0
 * implement required constraint for model attributes